Separate Chisel._ and chisel3._ tests to different specs
